refactor(lanes): split transform.py into transform/ package (move->squin, move->stack)
